# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CarKit API",
    "version": "1.0.0",
    "description": "CarKit API"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Serveur de développement"
    },
    {
      "url": "https://carkit.lazyy.fr",
      "description": "Serveur de production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Entrez votre token JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "message": {
            "type": "string",
            "example": "Error message"
          }
        }
      },
...
      "AuthResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "token": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        }
      },
...
      "VehicleInput": {
        "type": "object",
        "required": [
          "brand",
          "model",
          "customName",
          "motorization"
        ],
        "properties": {
          "brand": {
            "type": "string",
            "example": "Toyota"
          },
          "model": {
            "type": "string",
            "example": "Corolla"
          },
          "customName": {
            "type": "string",
            "example": "Ma voiture"
          },
          "motorization": {
            "type": "string",
            "enum": [
              "thermal",
              "hybrid",
              "electric"
            ],
            "example": "hybrid"
          },
          "imageUrl": {
            "type": "string",
            "format": "uri"
          },
          "year": {
            "type": "number",
            "minimum": 1900,
            "maximum": 2026
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/auth/apple": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Authentification avec Apple",
        "description": "Authentifie un utilisateur via Apple Sign In",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "identityToken"
                ],
                "properties": {
                  "identityToken": {
                    "type": "string",
                    "description": "Token d'identité Apple"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentification réussie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh-token/{refreshToken}": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Rafraîchir les tokens d'authentification",
        "description": "Génère un nouveau token et refreshToken à partir d'un refreshToken valide",
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "refreshToken",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Le refresh token à utiliser pour générer de nouveaux tokens"
          }
        ],
        "responses": {
          "200": {
            "description": "Tokens rafraîchis avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token invalide ou expiré",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Utilisateur non trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/google": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Authentification avec Google",
        "description": "Authentifie un utilisateur via Google Sign In",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "identityToken"
                ],
                "properties": {
                  "identityToken": {
                    "type": "string",
                    "description": "Token d'identité Google"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentification réussie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/vehicles": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Liste tous les véhicules",
        "description": "Récupère tous les véhicules de l'utilisateur connecté",
        "responses": {
          "200": {
            "description": "Liste des véhicules",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Vehicle"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Crée un nouveau véhicule",
        "description": "Ajoute un nouveau véhicule pour l'utilisateur connecté",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Véhicule créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/vehicles/{id}": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Récupère un véhicule par ID",
        "description": "Récupère les détails d'un véhicule spécifique",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID du véhicule"
          }
        ],
        "responses": {
          "200": {
            "description": "Détails du véhicule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              }
            }
          },
          "404": {
            "description": "Véhicule non trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Met à jour un véhicule",
        "description": "Modifie les informations d'un véhicule existant",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID du véhicule"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Véhicule mis à jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              }
            }
          },
          "404": {
            "description": "Véhicule non trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Supprime un véhicule",
        "description": "Supprime définitivement un véhicule et toutes ses données associées",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID du véhicule"
          }
        ],
        "responses": {
          "204": {
            "description": "Véhicule supprimé avec succès"
          },
          "404": {
            "description": "Véhicule non trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/vehicles/{vehicleId}/spendings": {
      "get": {
        "tags": [
          "Spendings"
        ],
        "summary": "Liste toutes les dépenses d'un véhicule",
        "description": "Récupère l'historique des dépenses d'un véhicule",
        "parameters": [
          {
            "in": "path",
            "name": "vehicleId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID du véhicule"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des dépenses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Spending"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Spendings"
        ],
        "summary": "Ajoute une nouvelle dépense",
        "description": "Enregistre une nouvelle dépense pour un véhicule",
        "parameters": [
          {
            "in": "path",
            "name": "vehicleId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID du véhicule"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpendingInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dépense créée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spending"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}